moved total solution to graph
