+++ src/graph/total_solution.py
