+++ src/solver.py
@@ -1,5 +1,5 @@
 from geometrics.ikosaeder import EPSILON
-from total_solution import Total_Solution
+from graph.total_solution import Total_Solution
 from greedy.greedysearch import GreedySearch
 from graph.graph import Graph
 import geometrics.seperation as seperator
